:art: Divide the library into modules
